+++ zoo-modules/form-modules/quantity-control-module/quantity-control.js
@@ -23,6 +23,10 @@ export default class QuantityControl extends HTMLElement {
 		const inputSlot = this.shadowRoot.querySelector('slot[name="input"]');
 		inputSlot.addEventListener('slotchange', () => {
 			this.input = inputSlot.assignedElements()[0];
+			this.input.addEventListener('invalid', () => this.setAttribute('invalid', ''));
+			this.input.addEventListener('change', e => {
+				e.target.checkValidity() ? this.removeAttribute('invalid') : this.setAttribute('invalid', '');
+			});
 			this.setInputWidth();
 		});